+++ app/containers/ActionSheet/Button.js
@@ -3,6 +3,5 @@ import { TouchableOpacity } from 'react-native';
 import { isAndroid } from '../../utils/deviceInfo';
 import Touch from '../../utils/touch';
 
-// For some reason react-native-gesture-handler isn't working on bottom sheet (iOS)
-const Button = isAndroid ? Touch : TouchableOpacity;
-export default Button;
+// Taken from https://github.com/rgommezz/react-native-scroll-bottom-sheet#touchables
+export const Button = isAndroid ? Touch : TouchableOpacity;
