+++ app/containers/MessageActions/index.js
@@ -13,7 +13,7 @@ import { getMessageTranslation } from '../message/utils';
 import { LISTENER } from '../Toast';
 import EventEmitter from '../../utils/events';
 import { showConfirmationAlert } from '../../utils/info';
-import { useActionSheet } from '../ActionSheet/Provider';
+import { useActionSheet } from '../ActionSheet';
 import Header, { HEADER_HEIGHT } from './Header';
 
 const MessageActions = React.memo(forwardRef(({
